Introduces compiler macro
