+++ Source/FieldSolver/Make.package
@@ -3,7 +3,9 @@ CEXE_headers += WarpX_FDTD.H
 CEXE_sources += WarpXPushFieldsEM.cpp
 ifeq ($(USE_PSATD),TRUE)
   include $(WARPX_HOME)/Source/FieldSolver/SpectralSolver/Make.package
-  include $(WARPX_HOME)/Source/FieldSolver/PicsarHybridSpectralSolver/Make.package
+  ifeq ($(USE_CUDA),FALSE)
+    include $(WARPX_HOME)/Source/FieldSolver/PicsarHybridSpectralSolver/Make.package
+  endif
 endif
 ifeq ($(USE_OPENBC_POISSON),TRUE)
   F90EXE_sources += openbc_poisson_solver.F90
