+++ Source/Make.WarpX
@@ -111,20 +111,20 @@ endif
    
 
 ifeq ($(USE_PSATD),TRUE)
-  FFTW_HOME ?= NOT_SET
-  ifneq ($(FFTW_HOME),NOT_SET)
-    VPATH_LOCATIONS += $(FFTW_HOME)/include
-    INCLUDE_LOCATIONS += $(FFTW_HOME)/include
-    LIBRARY_LOCATIONS += $(FFTW_HOME)/lib
-  endif
   USERSuffix := $(USERSuffix).PSATD
   DEFINES += -DWARPX_USE_PSATD
-  ifeq ($(USE_CUDA),TRUE)
-    DEFINES += -DFFTW -DCUDA_FFT=1 # PICSAR uses it
-    LIBRARIES += -lfftw3_mpi -lfftw3 -lfftw3_threads -lcufft
-  else
-    DEFINES += -DFFTW # PICSAR uses it
-    LIBRARIES += -lfftw3_mpi -lfftw3 -lfftw3_threads
+  ifeq ($(USE_CUDA),FALSE) # Running on CPU
+     # Use FFTW
+     LIBRARIES += -lfftw3_mpi -lfftw3 -lfftw3_threads
+     FFTW_HOME ?= NOT_SET
+     ifneq ($(FFTW_HOME),NOT_SET)
+       VPATH_LOCATIONS += $(FFTW_HOME)/include
+       INCLUDE_LOCATIONS += $(FFTW_HOME)/include
+       LIBRARY_LOCATIONS += $(FFTW_HOME)/lib
+     endif
+     # Compile with PICSAR's Hybrid-MPI PSATD
+     DEFINES += -DWARPX_USE_PSATD_HYBRID
+     DEFINES += -DFFTW # PICSAR uses it
   endif
 endif
